refactor(account): Moved persistense directory into infrastructure
