+++ module/account/src/Infrastructure/Persistence/Projector/User/DbalUserCreatedEventProjector.php
@@ -7,7 +7,7 @@
 
 declare(strict_types = 1);
 
-namespace Ergonode\Account\Persistence\Dbal\Projector\User;
+namespace Ergonode\Account\Infrastructure\Persistence\Projector\User;
 
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\DBALException;
@@ -16,7 +16,7 @@ use JMS\Serializer\SerializerInterface;
 
 /**
  */
-class UserCreatedEventProjector
+class DbalUserCreatedEventProjector
 {
     private const TABLE = 'users';
