+++ module/account/src/Infrastructure/Persistence/Query/DbalProfileQuery.php
@@ -7,12 +7,12 @@
 
 declare(strict_types = 1);
 
-namespace Ergonode\Account\Persistence\Dbal\Query;
+namespace Ergonode\Account\Infrastructure\Persistence\Query;
 
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\Query\QueryBuilder;
-use Ergonode\SharedKernel\Domain\Aggregate\UserId;
 use Ergonode\Account\Domain\Query\ProfileQueryInterface;
+use Ergonode\SharedKernel\Domain\Aggregate\UserId;
 
 /**
  */
