+++ module/account/src/Resources/config/services.yml
@@ -4,8 +4,8 @@ services:
         autoconfigure: true
         public: false
 
-    Ergonode\Account\Persistence\:
-        resource: '../../Persistence/*'
+    Ergonode\Account\Infrastructure\Persistence\:
+        resource: '../../Infrastructure/Persistence/*'
 
     Ergonode\Account\Application\:
         resource: '../../Application/*'
@@ -34,21 +34,21 @@ services:
         resource: '../../Infrastructure/Handler/*'
         tags: ['messenger.message_handler']
 
-    Ergonode\Account\Persistence\Dbal\Projector\:
-        resource: '../../Persistence/Dbal/Projector/*'
+    Ergonode\Account\Infrastructure\Persistence\Projector\:
+        resource: '../../Infrastructure/Persistence/Projector/*'
         tags: ['messenger.message_handler']
 
     Ergonode\Account\Infrastructure\JMS\Serializer\Handler\:
         resource: '../../Infrastructure/JMS/Serializer/Handler/*'
         tags: ['jms_serializer.subscribing_handler']
 
-    Ergonode\Account\Domain\Repository\UserRepositoryInterface: '@Ergonode\Account\Persistence\Dbal\Repository\DbalUserRepository'
-    Ergonode\Account\Domain\Repository\RoleRepositoryInterface: '@Ergonode\Account\Persistence\Dbal\Repository\DbalRoleRepository'
-    Ergonode\Account\Domain\Query\AccountQueryInterface: '@Ergonode\Account\Persistence\Dbal\Query\DbalAccountQuery'
-    Ergonode\Account\Domain\Query\ProfileQueryInterface: '@Ergonode\Account\Persistence\Dbal\Query\DbalProfileQuery'
-    Ergonode\Account\Domain\Query\PrivilegeQueryInterface: '@Ergonode\Account\Persistence\Dbal\Query\DbalPrivilegeQuery'
-    Ergonode\Account\Domain\Query\RoleQueryInterface: '@Ergonode\Account\Persistence\Dbal\Query\DbalRoleQuery'
-    Ergonode\Account\Domain\Query\UserQueryInterface: '@Ergonode\Account\Persistence\Dbal\Query\DbalUserQuery'
-    Ergonode\Account\Domain\Query\LogQueryInterface: '@Ergonode\Account\Persistence\Dbal\Query\DbalLogQuery'
+    Ergonode\Account\Domain\Repository\UserRepositoryInterface: '@Ergonode\Account\Infrastructure\Persistence\Repository\DbalUserRepository'
+    Ergonode\Account\Domain\Repository\RoleRepositoryInterface: '@Ergonode\Account\Infrastructure\Persistence\Repository\DbalRoleRepository'
+    Ergonode\Account\Domain\Query\AccountQueryInterface: '@Ergonode\Account\Infrastructure\Persistence\Query\DbalAccountQuery'
+    Ergonode\Account\Domain\Query\ProfileQueryInterface: '@Ergonode\Account\Infrastructure\Persistence\Query\DbalProfileQuery'
+    Ergonode\Account\Domain\Query\PrivilegeQueryInterface: '@Ergonode\Account\Infrastructure\Persistence\Query\DbalPrivilegeQuery'
+    Ergonode\Account\Domain\Query\RoleQueryInterface: '@Ergonode\Account\Infrastructure\Persistence\Query\DbalRoleQuery'
+    Ergonode\Account\Domain\Query\UserQueryInterface: '@Ergonode\Account\Infrastructure\Persistence\Query\DbalUserQuery'
+    Ergonode\Account\Domain\Query\LogQueryInterface: '@Ergonode\Account\Infrastructure\Persistence\Query\DbalLogQuery'
     Ergonode\Account\Infrastructure\Encoder\UserPasswordEncoderInterface: '@Ergonode\Account\Infrastructure\Encoder\DomainUserPasswordEncoder'
     Ergonode\Account\Infrastructure\Resolver\PrivilegeTypeResolverInterface: '@Ergonode\Account\Infrastructure\Resolver\PrivilegeTypeResolver'
